added master branch for conflict check, moved stock componets inside components
